Move OSMO runtime payload under scripts
Separate the workflow-construction task definition from the Python program embedded and executed inside the final OSMO task.
